requirements.txt
lxml
pillow
netcrawl
selenium